requirements.txt
PySide6>=6.6.0
opencv-python>=4.9.0
numpy>=1.26.0
python-dotenv>=1.0.0
openai>=1.30.0
djitellopy>=2.5.0
pygame>=2.5.0
colorama>=0.4.6
